#74 Add some unit tests and update dataset comparison
