+++ e2eRunner/src/main/scala/za/co/absa/hermes/e2eRunner/E2ERunnerJob.scala
@@ -15,150 +15,106 @@
 
 package za.co.absa.hermes.e2eRunner
 
-import com.typesafe.config.{Config, ConfigFactory, ConfigObject}
-import org.apache.spark.sql.SparkSession
-import za.co.absa.hermes.datasetComparison.cliUtils.{CliOptions, DataframeOptions}
-import za.co.absa.hermes.infoFileComparison.{InfoComparisonArguments, InfoFileComparisonJob, InfoFilesDifferException}
-import za.co.absa.hermes.datasetComparison.{DatasetComparisonException, DatasetComparisonJob}
-import za.co.absa.hermes.utils.HelperFunctions
+import java.io.File
 
-import scala.collection.JavaConversions._
-import scala.collection.mutable.ListBuffer
-import scala.sys.process._
 import scala.util.{Failure, Success, Try}
 
 object E2ERunnerJob {
-  private val conf: Config = ConfigFactory.load
-
-  private val sparkExecutable = conf.getString("e2e-runner.sparkSubmitExecutable")
-
-  private val dceJarPath = conf.getString("e2e-runner.dceJarPath")
-  private val dceStdName = conf.getString("e2e-runner.dceStdName")
-  private val dceConfName = conf.getString("e2e-runner.dceConfName")
-
-  private val stdClass = conf.getString("e2e-runner.stdClass")
-  private val confClass = conf.getString("e2e-runner.confClass")
-
-  private val stdOutputPath = conf.getString("e2e-runner.stdPath")
-  private val confOutputPath = conf.getString("e2e-runner.confPath")
-
-  private val sparkSubmitOptions = conf.getString("e2e-runner.sparkSubmitOptions")
-
-  private val confExtraJavaOpts = conf.getObject("e2e-runner.sparkConf.extraJavaOptions")
-
-  private val stdFullJarPath = s"$dceJarPath/$dceStdName"
-  private val confFullJarPath = s"$dceJarPath/$dceConfName"
-
-  private val stdParamsOverride = Try(conf.getString("e2e-runner.stdParams")).filter(_.nonEmpty).toOption
-  private val confParamsOverride = Try(conf.getString("e2e-runner.confParams")).filter(_.nonEmpty).toOption
-
-  private def runBashCmd(bashCmd: String): String = {
-    (s"echo $bashCmd" #| "bash").!!
-  }
-
-  private def createSparkConfExtraJavaOptions(confObject: ConfigObject, stdPathHDFS: String): String = {
-    val extraJavaConfPrefix = "--conf 'spark.driver.extraJavaOptions="
-    val confWithoutLastApostrophe = confObject.unwrapped().foldLeft(extraJavaConfPrefix){ (acc, confOption) =>
-      val (key, value) = confOption
-      s"""$acc-D$key="${value}" """
-    }
-
-    s"""$confWithoutLastApostrophe -Dstandardized.hdfs.path="$stdPathHDFS"'"""
-  }
-
-  private def getDatasetComparisonConfig(cmd: E2ERunnerConfig, pathHDFS: String): CliOptions = {
-    val keys = if (cmd.keys.isDefined) { cmd.keys.get.toSet } else { Set.empty[String] }
-
-    CliOptions(
-      DataframeOptions("parquet", Map.empty[String, String], s"/ref$pathHDFS"),
-      DataframeOptions("parquet", Map.empty[String, String], pathHDFS),
-      Some(DataframeOptions("parquet", Map.empty[String, String], s"/comp$pathHDFS")),
-      keys,
-      "ran from E2E Runner"
-    )
-  }
-
-  private def getInfoComparisonConfig(cmd: E2ERunnerConfig, pathHDFS: String): InfoComparisonArguments = {
-    InfoComparisonArguments(
-      newPath = s"$pathHDFS/_INFO",
-      refPath = s"/ref$pathHDFS/_INFO",
-      outPath = s"/comp/info_file$pathHDFS/_INFO"
-    )
-  }
+  /**
+   * To always add plugins from this project we add this JAR file's path to the list of class paths
+   */
+  private val locationToThisClass = new File(
+    this.getClass.getProtectionDomain.getCodeSource.getLocation.toURI
+  ).getPath
 
   def main(args: Array[String]): Unit = {
-    val cmd = E2ERunnerConfig.getCmdLineArguments(args, stdParamsOverride, confParamsOverride) match {
+    val cmd = E2ERunnerConfig.getCmdLineArguments(args) match {
       case Success(value) => value
       case Failure(exception) => throw exception
     }
 
-    implicit val sparkSession: SparkSession = SparkSession.builder()
-      .appName(s"End2End test of ${cmd.datasetName} (v.${cmd.datasetVersion}) " +
-        s"${cmd.reportVersion} v.${cmd.reportVersion}")
-      .getOrCreate()
+    val classpath = (List(locationToThisClass) ++ cmd.jarPath).map(new File(_))
+    scribe.info(
+      s"""Plugin classes will be loaded from these paths:
+         |${classpath.mkString("\n")}""".stripMargin)
 
-    val stdPathHDFS = HelperFunctions.richFormat(stdOutputPath, cmd.getDatasetMap)
-    val confPathHDFS = HelperFunctions.richFormat(confOutputPath, cmd.getDatasetMap)
+    val pluginDefinitions: PluginDefinitions = PluginDefinitions(classpath)
+    val pluginNames = pluginDefinitions.getPluginNames
+    scribe.info(
+      s"""Loaded plugins are:
+         |${pluginNames.mkString("\n")}""".stripMargin)
 
-    val stdParams = if (stdParamsOverride.isDefined) stdParamsOverride.get else cmd.getStdParams
-    val confParams = if (confParamsOverride.isDefined) confParamsOverride.get else cmd.getConfParams
+    val testDefinitions = TestDefinition.fromFile(cmd.testDefinition)
+    scribe.info(s"Loaded ${testDefinitions.size} test definitions")
 
-    val extraJavaOpts = createSparkConfExtraJavaOptions(confExtraJavaOpts, stdPathHDFS)
+    val pluginsExpectedToUse = testDefinitions.getPluginNames
+    validatePluginsToBeUsed(pluginNames, pluginsExpectedToUse)
 
-    val sparkJobPrefix = s"$sparkExecutable $sparkSubmitOptions --class"
-    val standardization = s"$sparkJobPrefix $stdClass $extraJavaOpts $stdFullJarPath $stdParams"
-    val conformance = s"$sparkJobPrefix $confClass $extraJavaOpts $confFullJarPath $confParams"
+    scribe.info("Running tests")
+    val results = runTests(testDefinitions, pluginDefinitions, cmd.failfast)
 
-    val errAccumulator = new ListBuffer[String]()
+    scribe.info("##################################################")
+    scribe.info("Invoking logging of test results")
+    scribe.info("##################################################")
+    results.foreach({ result =>
+      result.logResult()
+      scribe.info("##################################################")
+    })
+  }
 
-    scribe.info(s"Standardization started - $standardization")
-    val (stdTime, _) = HelperFunctions.calculateTime { runBashCmd(standardization) }
-    scribe.info("Standardization Passed")
+  /**
+   * Runs test definitions one by one
+   * @param testDefinitions Tests to be run
+   * @param pluginDefinitions Plugins available
+   * @param failFast Should the tests fail if one of them throws exception
+   * @return Returns a sequence of PluginResults
+   */
+  def runTests(testDefinitions: TestDefinitions,
+               pluginDefinitions: PluginDefinitions,
+               failFast: Boolean = false): Seq[PluginResult] = {
+    testDefinitions.foldLeftWithIndex {
+      case (acc, TestDefinitionWithOrder(td, order)) =>
+        scribe.info(s"Running ${td.name}")
+        val plugin: Plugin = pluginDefinitions.getPlugin(td.pluginName)
 
-    scribe.info("Standardization Comparison started")
-    scribe.info("Standardization Comparison of Dataset")
-    try {
-      DatasetComparisonJob.execute(getDatasetComparisonConfig(cmd, stdPathHDFS))
-    } catch {
-      case e: DatasetComparisonException => errAccumulator.add(e.getMessage)
+        val tryExecution = if (canTestProceed(td, acc)) {
+          tryExecute(td, order, plugin)
+        } else {
+          Failure(throw DependeeFailed(td.name, td.dependsOn.get))
+        }
+
+        tryExecution match {
+          case Success(value) => acc :+ value
+          case Failure(exception) if !failFast =>  acc :+ FailedPluginResult(td.args, exception, order, td.name)
+          case Failure(exception) if failFast =>  throw TestFailedWithFailFastOn(td.name, order, exception)
+        }
     }
+  }
 
-    scribe.info("Standardization Comparison of INFO file")
-    try {
-      InfoFileComparisonJob.execute(getInfoComparisonConfig(cmd, stdPathHDFS))
-    } catch {
-      case e: InfoFilesDifferException => errAccumulator.add(e.getMessage)
+  private def canTestProceed(td: TestDefinition, previousResults: Seq[PluginResult]): Boolean = {
+    td.dependsOn.forall { dependerName =>
+      previousResults.find({ previousResult =>
+        dependerName.equalsIgnoreCase(previousResult.getTestName)
+      }).forall(_.testPassed)
     }
+  }
 
-    scribe.info(s"Conformance started - $conformance")
-    val (confTime, _) = HelperFunctions.calculateTime { runBashCmd(conformance) }
-    scribe.info("Conformance Passed")
+  private def tryExecute(td: TestDefinition,
+                         testOrder: Int,
+                         plugin: Plugin): Try[PluginResult] = Try {
+    val result: PluginResult = plugin.performAction(td.args, testOrder, td.name)
+    if (td.writeArgs.isDefined) result.write(td.writeArgs.get)
+    result.logResult()
+    result
+  }
 
-    scribe.info("Conformance Comparison started")
-    scribe.info("Conformance Comparison of Dataset")
-    try {
-      DatasetComparisonJob.execute(getDatasetComparisonConfig(cmd, confPathHDFS))
-    } catch {
-      case e: DatasetComparisonException => errAccumulator.add(e.getMessage)
-    }
-
-    scribe.info("Conformance Comparison of INFO file")
-    try {
-      InfoFileComparisonJob.execute(getInfoComparisonConfig(cmd, confPathHDFS))
-    } catch {
-      case e: InfoFilesDifferException => errAccumulator.add(e.getMessage)
-    }
-
-    val humanReadableStdTime = HelperFunctions.prettyPrintElapsedTime(stdTime)
-    val humanReadableConfTime = HelperFunctions.prettyPrintElapsedTime(confTime)
-    scribe.info(s"Standardization and Conformance passed. It took them $humanReadableStdTime and " +
-      s"$humanReadableConfTime respectively")
-
-    if (errAccumulator.nonEmpty) {
-      throw E2EComparisonException(errAccumulator.mkString("\n"))
-    } else {
-      scribe.info("All comparisons have passed")
-    }
-
+  /**
+   * Validates if all the needed plugins are there
+   * @param pluginNames Names of loaded plugins
+   * @param pluginsExpectedToUse Names of plugins from all test definitions
+   */
+  private def validatePluginsToBeUsed(pluginNames: Set[String], pluginsExpectedToUse: Set[String]): Unit = {
+    if (pluginsExpectedToUse.diff(pluginNames).nonEmpty)
+      throw NotAllRequestedPluginsDefined(pluginNames.diff(pluginsExpectedToUse))
   }
 }
