+++ e2eRunner/src/test/scala/za/co/absa/hermes/e2eRunner/E2ERunnerJobSuite.scala
@@ -3,7 +3,7 @@ package za.co.absa.hermes.e2eRunner
 import org.scalatest.FunSuite
 import za.co.absa.atum.utils.SparkTestBase
 
-class E2ERunnerJobExperimentalSuite extends FunSuite with SparkTestBase {
+class E2ERunnerJobSuite extends FunSuite with SparkTestBase {
 //  test("alfa") {
 //    E2ERunnerJobExperimental.main(Array(getClass.getResource("/alfa.json").toString))
 //  }
